Modularize, use generators instead of loop
