+++ lambda/kinesis_logging_utils/misc.py
@@ -1,0 +1,6 @@
+from typing import List
+
+
+def split_list(lst: list, size: int) -> List[list]:
+    for i in range(0, len(lst), size):
+        yield lst[i:i + size]
